refactor(notifications): decompose NotificationSection into cards
